Create rcolumsfinderror.py
This one does what read columns does and then it takes one set of readings and uses it to interpolate for the second set. eg. it will take vcalibrated and Icalibrated and then find the interpolated values of current from the second dataset for the vcalibrated values. It will also print the error of the interpolated values and the values of the calibrated dataset